feat: refactor credentials forms with reusable components and error handling
