+++ ui/components/providers/workflow/forms/index.ts
@@ -1,6 +1,6 @@
+export * from "./add-via-credentials-form";
 export * from "./connect-account-form";
 export * from "./test-connection-form";
 export * from "./update-via-credentials-form";
 export * from "./update-via-role-form";
-export * from "./via-credentials-form";
 export * from "./via-role-form";
