+++ ui/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from "./use-credentials-form";
 export * from "./use-form-server-errors";
 export * from "./use-local-storage";
 export * from "./use-sidebar";
